.history/api/serializers_20210212150231.py
from rest_framework import serializers
from .models import *
from .models import User

class RegisterSerializer(serializers.Serializer):
    username = serializers.CharField(max_length=255)
    password = serializers.CharField(max_length=255)