requirements.txt
asgiref==3.3.1
Django==3.1.6
djangorestframework==3.12.2
mccabe==0.6.1
Pillow==8.1.0
psycopg2-binary==2.8.6
pycodestyle==2.6.0
pydocstyle==5.1.1
pyflakes==2.2.0
pylama==7.7.1
pytz==2021.1
snowballstemmer==2.1.0
sqlparse==0.4.1
